backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
passlib[bcrypt]
PyJWT
python-multipart
mysql-connector-python
